requirements.txt
# SOLARIS-X Advanced ML Requirements - CPU Optimized & Python 3.13 Compatible
# =============================================================================

# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyarrow>=13.0.0

# Machine Learning Frameworks (CPU Optimized)
tensorflow-cpu>=2.13.0
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# High-Performance Computing (CPU Focus)
numba>=0.58.0
joblib>=1.3.0
threadpoolctl>=3.2.0

# Model Training & Hyperparameter Optimization
optuna>=3.0.0
hyperopt>=0.2.7
bayesian-optimization>=1.4.0

# Model Evaluation & Interpretability
shap>=0.44.0
lime>=0.2.0.1
eli5>=0.13.0

# Advanced Metrics & Validation
imbalanced-learn>=0.11.0
sklearn-pandas>=2.2.0
category-encoders>=2.6.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0

# Progress & System Monitoring
tqdm>=4.65.0
psutil>=5.9.0
memory_profiler>=0.61.0

# Model Persistence & Deployment (Python 3.13 Compatible)
dill>=0.3.7
cloudpickle>=2.2.0

# Development & Analysis
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Utilities & Data Handling
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3

# Statistical Analysis & Time Series
scipy>=1.11.0
statsmodels>=0.14.0
